Header cleanup, root fix.
