+++ src/cymoca/ast/ast.h
@@ -2,7 +2,6 @@
 #define CYMOCA_AST_AST_H_
 
 #include "condition/condition.h"
-#include "declarations.h"
 #include "element/element.h"
 #include "equation/equation.h"
 #include "expression/expression.h"
