+++ src/cymoca/ast/listener/lisp.h
@@ -24,6 +24,7 @@ class Lisp : public listener::Base {
   void enter(condition::LessThan &) override { m_ss << "<"; }
   void enter(condition::LessThanOrEqual &) override { m_ss << "<="; }
   void enter(model::Class &) override { m_ss << "class"; }
+  void enter(model::ElementDict &) override { m_ss << "elem dict"; }
   void enter(equation::List &) override { m_ss << "eq list"; }
   void enter(equation::If &) override { m_ss << "if"; }
   void enter(equation::Simple &) override { m_ss << "="; }
