+++ src/cymoca/ast/statement/statement.h
@@ -6,11 +6,6 @@
 namespace cymoca::ast::statement {
 
 /**
- * Base statment type.
- */
-class Base : public INode {};
-
-/**
  * A block of statements with a condition
  * controlling the block execution.
  */
